packages/@aws-prototyping/fraud-detection/app/api.requirements.txt
boto3
fastapi
pydantic
mangum
uvicorn
python-multipart
python-dotenv
Pillow
albumentations
tqdm
langchain
sagemaker
boto3
opensearch-py
requests-aws4auth
langchain
langchain_aws
langchain-community
pandas
rich
opencv-python-headless
numpy
tqdm
piexif
exif
geopy
pydeck
google-search-results
serpapi
streamlit-cognito-auth
fastapi
altair
timm
loguru
scipy
sqlitedict
pylint
annoy
sentence-transformers
elasticsearch